@@ -43,7 +43,8 @@ dependencies:
   flutter_keyboard_visibility: ^5.3.0
   emoji_picker_flutter: ^1.4.0
   should_rebuild: ^1.0.1
-  flutter_openim_sdk: ^3.4.0+2
+  flutter_openim_sdk: ^3.5.0-rc.1+1
+  common_utils: 2.1.0
 #  flutter_openim_sdk:
 ##    path: ../flutter_openim_sdk
 #    git:
