@@ -1,146 +1,115 @@
 import 'dart:io';
 
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_openim_widget/flutter_openim_widget.dart';
+import 'package:flutter_openim_widget/src/res/images.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+import '../flutter_openim_widget.dart';
+import 'chat/chat_progress_view.dart';
 
 class ChatVideoView extends StatefulWidget {
-  final String? videoPath;
-  final String? videoUrl;
-  final String? snapshotUrl;
-  final String? snapshotPath;
-  final double? width;
-  final double? height;
-  final double widgetWidth;
-  final bool isReceived;
-  final String msgId;
-  final Stream<MsgStreamEv<int>>? msgSenProgressStream;
-  final int initMsgSendProgress;
-  final int index;
-  final Stream<int>? clickStream;
-
   const ChatVideoView({
     Key? key,
-    required this.msgId,
-    required this.isReceived,
-    required this.index,
-    this.clickStream,
-    this.snapshotPath,
-    this.snapshotUrl,
-    this.videoPath,
-    this.videoUrl,
-    this.width,
-    this.height,
-    this.widgetWidth = 100,
-    this.msgSenProgressStream,
-    this.initMsgSendProgress = 100,
+    required this.message,
+    required this.isISend,
+    this.sendProgressStream,
   }) : super(key: key);
+  final bool isISend;
+  final Stream<MsgStreamEv<int>>? sendProgressStream;
+  final Message message;
 
   @override
-  _ChatVideoViewState createState() => _ChatVideoViewState();
+  State<ChatVideoView> createState() => _ChatVideoViewState();
 }
 
 class _ChatVideoViewState extends State<ChatVideoView> {
   late double _trulyWidth;
   late double _trulyHeight;
-  String? snapshotUrl;
-  String? snapshotPath;
-  String? url;
-  String? path;
+  String? _snapshotUrl;
+  String? _snapshotPath;
+  Widget? _child;
 
+  Message get _message => widget.message;
   @override
   void initState() {
-    path = widget.videoPath;
-    url = widget.videoUrl;
-    snapshotUrl = widget.snapshotUrl;
-    snapshotPath = widget.snapshotPath;
+    final video = _message.videoElem;
+    _snapshotUrl = video?.snapshotUrl;
+    _snapshotPath = video?.snapshotPath;
 
-    var w = widget.width ?? 1.0;
-    var h = widget.height ?? 1.0;
+    var w = video?.snapshotWidth?.toDouble() ?? 1.0;
+    var h = video?.snapshotHeight?.toDouble() ?? 1.0;
 
-    _trulyWidth = widget.widgetWidth;
+    _trulyWidth = pictureWidth;
     _trulyHeight = _trulyWidth * h / w;
-    /*if (widget.widgetWidth > w) {
-      _trulyWidth = w;
-      _trulyHeight = h;
-    } else {
-      _trulyWidth = widget.widgetWidth;
-      _trulyHeight = _trulyWidth * h / w;
-    }*/
 
-    /*widget.clickStream?.listen((i) {
-      if (!mounted) return;
-      if (_isClickedLocation(i)) {
-        if (null != widget.onClickVideo) {
-          widget.onClickVideo!(url, path);
-        }
-      }
-    });*/
+    _createThumbView();
     super.initState();
   }
 
-  bool _isClickedLocation(i) => i == widget.index;
+  Future<bool> _checkingPath() async {
+    final valid = IMUtils.isNotNullEmptyStr(_snapshotPath) &&
+        // await Permissions.checkStorage() &&
+        await File(_snapshotPath!).exists();
+    _message.exMap['validPath_$_snapshotPath'] = valid;
+    return valid;
+  }
 
-  Widget _buildThumbView() {
-    if (widget.isReceived) {
-      if (null != snapshotUrl && snapshotUrl!.isNotEmpty) {
-        return ImageUtil.networkImage(
-          url: snapshotUrl!,
-          width: _trulyWidth,
-          height: _trulyHeight,
-          fit: BoxFit.fitWidth,
-        );
-      }
-    } else {
-      if (null != snapshotPath &&
-          snapshotPath!.isNotEmpty &&
-          File(snapshotPath!).existsSync()) {
-        return Image(
-          image: FileImage(File(snapshotPath!)),
-          height: _trulyHeight,
-          width: _trulyWidth,
-          fit: BoxFit.fitWidth,
-          errorBuilder: (_, error, stack) => _errorIcon(),
-        );
-      } else {
-        if (null != snapshotUrl && snapshotUrl!.isNotEmpty) {
-          return ImageUtil.networkImage(
-            url: snapshotUrl!,
-            width: _trulyWidth,
-            height: _trulyHeight,
-            fit: BoxFit.fitWidth,
-          );
-        }
-      }
+  bool? get isValidPath => _message.exMap['validPath_$_snapshotPath'];
+
+  _createThumbView() async {
+    if (widget.isISend &&
+        (isValidPath == true || isValidPath == null && await _checkingPath())) {
+      _child = ImageUtil.fileImage(
+        File(_snapshotPath!),
+        height: _trulyHeight,
+        width: _trulyWidth,
+        fit: BoxFit.fitWidth,
+      );
+    } else if (IMUtils.isNotNullEmptyStr(_snapshotUrl)) {
+      _child = ImageUtil.networkImage(
+        url: _snapshotUrl!,
+        width: _trulyWidth,
+        height: _trulyHeight,
+        fit: BoxFit.fitWidth,
+      );
     }
-    return Container(width: _trulyWidth, height: _trulyHeight);
+    if (null != _child) {
+      if (!mounted) return;
+      setState(() {});
+    }
   }
 
   @override
-  Widget build(BuildContext context) {
-    return Hero(
-        tag: widget.msgId,
-        child: Container(
-          width: _trulyWidth,
-          height: _trulyHeight,
-          // color: Color(0xFFB3D7FF),
-          child: Stack(
-            alignment: Alignment.center,
-            children: [
-              _buildThumbView(),
-              ImageUtil.play(),
-              ChatSendProgressView(
-                height: _trulyHeight,
-                width: _trulyWidth,
-                msgId: widget.msgId,
-                stream: widget.msgSenProgressStream,
-                initProgress: widget.initMsgSendProgress,
-              ),
-            ],
+  Widget build(BuildContext context) => ClipRRect(
+    borderRadius: borderRadius(widget.isISend),
+    child: SizedBox(
+      width: _trulyWidth,
+      height: _trulyHeight,
+      child: Stack(
+        alignment: Alignment.center,
+        children: [
+          if (null != _child) _child!,
+          ImageRes.videoPause.toImage
+            ..width = 40.w
+            ..height = 40.h,
+          ChatProgressView(
+            height: _trulyHeight,
+            width: _trulyWidth,
+            id: _message.clientMsgID!,
+            stream: widget.sendProgressStream,
+            isISend: widget.isISend,
+            type: ProgressType.video,
           ),
-        ));
-  }
-
-  Widget _errorIcon() =>
-      ImageUtil.error(width: _trulyWidth, height: _trulyHeight);
+          if (null != _message.videoElem?.duration)
+            Positioned(
+              bottom: 2.h,
+              right: 3.w,
+              child:
+              IMUtils.seconds2HMS(_message.videoElem!.duration!).toText
+                ..style = Styles.ts_FFFFFF_12sp,
+            ),
+        ],
+      ),
+    ),
+  );
 }
